Add EntityRepository and rename files
