+++ src/backends/Query.ts
@@ -6,8 +6,8 @@ import { BackendInterface } from './AbstractAdapter'
 import { Core } from '../Core'
 import { DataObjectClass, ObjectUri } from '../components'
 import { Persisted } from '../components/types/Persisted'
-import { BaseObject } from '../components/BaseObjectProperties'
-import { BaseObjectCore } from '../components/BaseObject'
+import { BaseObject } from '../components/BaseObject'
+import { BaseObjectCore } from '../components/BaseObjectCore'
 
 export const AS_OBJECTURIS = 'objectUris'
 export const AS_DATAOBJECTS = 'dataObjects'
