+++ src/components/BaseRepository.ts
@@ -3,12 +3,12 @@ import * as statuses from '../statuses'
 import { BackendInterface } from '../backends/AbstractAdapter'
 import { Query } from '../backends/Query'
 import { GoneError, NotFoundError } from '../common/ResourcesErrors'
-import { BaseObjectCore } from './BaseObject'
+import { BaseObjectCore } from './BaseObjectCore'
 import { DataObject } from './DataObject'
 import { DataObjectClass } from './types/DataObjectClass'
 import Payload, { Meta } from './types/Payload'
 import RepositoryClass from './types/RepositoryClass'
-import { BaseObject } from './BaseObjectProperties'
+import { BaseObject } from './BaseObject'
 import { Persisted } from './types/Persisted'
 import { Proxy } from './types/ProxyConstructor'
