+++ src/components/index.ts
@@ -2,8 +2,8 @@ import { DataObject, DataObjectParams } from './DataObject'
 import { DataObjectClass } from './types/DataObjectClass'
 import { ObjectUri } from './ObjectUri'
 import { AbstractObject } from './AbstractObject'
-import { BaseObjectCore } from './BaseObject'
-import { BaseObjectProperties } from './BaseObjectProperties'
+import { BaseObjectCore } from './BaseObjectCore'
+import { BaseObjectProperties } from './BaseObject'
 import { UserCore, UserProperties } from './User'
 import { Entity } from './Entity'
