+++ src/components/types/RepositoryClass.ts
@@ -1,7 +1,7 @@
 import { BackendInterface } from '../../backends/AbstractAdapter'
 import { Query } from '../../backends/Query'
-import { BaseObjectCore } from '../BaseObject'
-import { BaseObject } from '../BaseObjectProperties'
+import { BaseObjectCore } from '../BaseObjectCore'
+import { BaseObject } from '../BaseObject'
 import { Proxy } from './ProxyConstructor'
 import Payload from './Payload'
 import { Persisted } from './Persisted'
